feat: use status field in crd
we're also using a properly typed client now, no more json mess.
the server can now assign the address to the client however it wants
it should also add itself as a peer.

the client that created the object can watch it and see it get attributed
an address and generate it's config from that.